#127664# $CCNUMVER from dmake to configure
Move the calculation of $CCNUMVER and some other variables from
main/solenv/inc/tg_compv.mk, where it is only usable by dmake, to
configure, where it can be used by both dmake and gbuild. This is
a requirement to upstream some compiler bug workaround patches from
the FreeBSD port.

A bit of logic from set_soenv is also moved into configure.  A bunch
more should probably be moved so that the configuration logic is 
not spread across so many different places, but that can wait.
Something else to consider is that it would be nice to use a different
value of $COM for Apple's clang, maybe "ACLANG" or "APPLECLANG"
since it  has a different version numbering scheme that the open-source
version of clang and having a unique identifier would simplify
version checking when applying compiler bug workarounds.

Note: I think the old value of $CCNUMVER on the Mac is wrong.  It 
should look something like 000800010000 or 000700030000, depending
on the installed version.

Change -DCPPU_ENV on the Mac from $(COMID) to $(COMNAME) for
consistency with the dmake side.  It shouldn't make a difference
in practice since both have the same value on the Mac.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1826296 13f79535-47bb-0310-9956-ffa450edef68