python: rename pcbuild to msvs9 in patch names to avoid confusion.
Since r1759046, which updated python to version 2.7.12, we are using
the legacy buildfiles which upstream moved from PCbuild to PC/VS9.0.

We kept the old name for the patches to preserve the history in the
repository but since there is a good chance we may update the required
compiler, and use the updated build files, rename them properly to avoid
confusion.

Consistently use $(PYVERSION) while here.

Note this is untested: being a simple rename there should be no problem.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1826694 13f79535-47bb-0310-9956-ffa450edef68