Fix potential memory alignment issues on X86_64.
Clang version 4.0 and newer uses SSE instructions that require
16-byte alignment to zero memory allocated using the C++ new operator.
The internal memory allocator does not understand anything larger
than 8-byte alignment.  Modify it to be capable of doing 16-byte
alignment when necessary.

There is also a debug layer beneath the C++ new and delete operators
that is enabled by the --enable-debug configure option.  This layer
adds 8 to the requested size of any allocations before calling the
underlying allocator, adds a known signature to the start of the
memory block, and then adds an 8 byte offset to the pointer before
returning it to the caller.  The delete operator basically does the
reverse, checking for the proper signature.  Modify this code so
that it adds and subtracts a 16-byte offset on X86_64 so that a
properly aligned block from the underlying allocator does not cause
new to return a misaligned pointer.

Modify set_soenv.in so that it always requests 16-byte alignment
on X86_64 so that the ABI is the same independent of the toolchain.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1830406 13f79535-47bb-0310-9956-ffa450edef68