Drop references and the requirement for the StAX API.
The StAX API was a requirement to build saxon however, long ago StAX
became part of the Java API[1], so in practice we are not using StAX 
as a separate tarball at all.

No functional change intended.

Huge thanks to AdfinisSygroup that maintained the opengrok service for so 
long. These type of changes would have been much more difficult without 
the service.

[1]
https://docs.oracle.com/javase/tutorial/jaxp/stax/why.html



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1832747 13f79535-47bb-0310-9956-ffa450edef68