Move smoketestdoc to the top-level test directory. This is where all integration tests are meant to be; a binary copy of the smoketest was there already.
Delete the smoketestoo_native module whose only purpose was to run
the smoketest, something that is done in the Java bvt test now.
Delete the old smoketest binary copy in test/ and use this one built
from source.
Build an .odt file for the smoketest instead of the legacy .sxw.

Note that we now require main/ built to build test/, as
the Ant build files required to build smoketest's Java component
are in main/.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1838268 13f79535-47bb-0310-9956-ffa450edef68