Merge r1748497, r1758150, r1839130 from trunk:
#i126893#: Upgrade bundled libxml2 version to 2.9.3 and libxslt to 1.1.28

These security vulnerabilities have been fixed since the release of
libxml2-2.7.8 which is bundled with OpenOffice:
    CVE-2011-3202
    CVE-2011-3919
    CVE-2013-0338
    CVE-2013-0339
    CVE-2013-2877
    CVE-2014-0191
    CVE-2014-3660
    CVE-2015-1819
    CVE-2015-5312
    CVE-2015-7497
    CVE-2015-7498
    CVE-2015-7499
    CVE-2015-7500
    CVE-2015-7941
    CVE-2015-7942
    CVE-2015-8035
    CVE-2015-8241
    CVE-2015-8242

This vulnerability has been fixed since the release of libxslt-1.1.26:
    CVE-2015-7995
    
Upgrade libxml2 to version 2.9.3 which has no publicly disclosed
vulnerabilities at this time.
    
Rebase libxml2-configure.patch to the new version of libxml2.  The
freebsd-elf change to ltmain.sh is no longer necessary and is eliminated
from the patch.  The fixes in libxml2-fixes.patch were either fixed
upstream or don't seem to apply anymore, so this patch file was deleted.
The fixes in libxml2-testapi.patch and libxml2-runtest.patch are in now
in the upstream source, so these patch files have been deleted.
The libxml2-mingw.patch and Solaris-specific libxml2-global-symbols.patch
were not updated and are disconnected from the build.  Several of the
fixes in libxml2-long-path.patch are now fixed upstream.
    
Upgrade libxslt to version 1.1.28 and add libxslt-CVE-2015-7995.patch,
which is imported from the FreeBSD port and which appears to have been
cherry picked from upstream.

Rebase libxslt-configure.patch to the new version of libxslt, with
the libtool-related changes coming from the libxslt port to FreeBSD.
The fixes in libxslt-bsd.patch are now present in the upstream
source, so this patch was deleted.

Tested by:	kschenk


Upgrade bundled libxml2 from version 2.9.3 to version 2.9.4 to fix:
	CVE-2016-1762
	CVE-2016-1833
	CVE-2016-1834
	CVE-2016-1835
	CVE-2016-1836
	CVE-2016-1837
	CVE-2016-1838
	CVE-2016-1839
	CVE-2016-1840
	CVE-2016-3627
	CVE-2016-3705
	CVE-2016-4449
	CVE-2016-4483
It is not known whether any of these affect OpenOffice.
 
Import a post-2.9.4 patch from the upstream git repo to fix a build
problem on systems with pre-C99 compilers.



Upgrade bundled libxml2 to version 2.9.8 and libxslt to version
1.1.32.  Some of the patches for the previous version of libxml2
have been picked up upstream and are no longer needed.  One new
patch is needed for clean builds on Windows, so cherrypick it from
upstream.


Submitted by: truckman
Reviewed by: jim


git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/AOO416@1839349 13f79535-47bb-0310-9956-ffa450edef68