i127875 - Merge r1837212 from trunk
Fix SwCntntFrm::CalcLowers() so that it visits all on-page objects.
 SwCntntFrm::CalcLowers() walks the object tree walk and it assumes
that the objects visited will be at a monotonically increasing
distance from the top of the page, so it terminates the walk once 
it encounters the first object below the bottom of the page.  Even
though the objects are vertically sorted in each branch of the tree,
there is no guarantee that the order of the objects visited in the
depth-first walk will be perfectly sorted vertically, so the walk
may terminate before visiting all of the objects on the page.  Fix
this problem by pruning branches when they cross the lower page
boundary and only terminating the walk once there are no longer any
paths forward.

Submitted by: truckman

git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/AOO416@1843573 13f79535-47bb-0310-9956-ffa450edef68