Re-commit r1847839 and r1847840, with future patches on the way that will get the module building in gbuild and using source-level attributes to control symbol visibility. The original commit was worded as:
Change default symbol visibility to hidden in main/cppuhelper, and export
symbols using source-level attributes (SAL_DLLPUBLIC_EXPORT) instead.

Stop using the .map file for Windows, and rely only on the above to control
symbol visibility.

This is not a perfect solution, as the exported symbols lack some typeinfos
compared to before, but it doesn't seem to break anything, and a full symbol
audit needs to be done later to fix other modules anyway.

Also added some files to get it to build with gbuild, which are currently
unused, as the gbuild build seems to break in main/i18npool.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1848075 13f79535-47bb-0310-9956-ffa450edef68