Merge r1851443, r1851449 from trunk:
Use fakeroot instead of our custom getuid.c LD_PRELOAD hack if available when creating Debian packages. This is because epm and dpkg (et.al.) need to be fooled into thinking that root is creating these packages for the correct permissions.

As long as we have the full path, use it

Reviewed by: jim

git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/AOO42X@1851450 13f79535-47bb-0310-9956-ffa450edef68