Add a preliminary Windows AMD64 UNO-C++ bridge.
Largely based on a mixture of the Linux64 and Win32 bridges,
with significant AMD64 assembly language also used, it was
challenging to develop but luckily not too long.

Some links and stacks/register diagrams have been left
in the code for easy reference.

It compiles and links, but how well it works remains to be tested. At
least it's no longer holding up the Win64 port. The exception handling
is the least certain, although stack alignments for the assembly language
could also have issues.

Also patched gbuild to use the 64 bit MASM for building assembly language
files on Win64.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1852196 13f79535-47bb-0310-9956-ffa450edef68