Upgrade bundled expat to 2.2.9.
Delete expat-winapi.patch which was needed to get linkage on Windows
to work with the expat static libraries.  Instead add -DXML_STATIC
to the compiler flags when building the modules that link to the
expat static libraries to accomplish the same thing, but in a
supported way.

Don't build the expat man pages since we don't have the necessary
build dependencies and we don't need them anyway.