Fix incorrect gbuild usage that can cause intermittent parallel build failures.
There was some misuse of gbuild that was causing make rules to be added
that had solver/*/*/workdir/ExternalHeaders/Library and
solver/*/*/workdir/ExternalHeaders/Library listed as dependencies.
If these directories did not exist, then a pattern match rule intended
to create files under the
solver/*/*/workdir/ExternalHeaders/{Library,StaticLibrary} directories
would be triggered.  The recipe for that rule would then
  mkdir -p solver/*/*/workdir/ExternalHeaders
and then
  touch solver/*/*/workdir/ExternalHeaders/{Library,StaticLibrary}
creating it as a plain file.  A subsequent make rule would
unconditionally
  mkdir -p solver/*/*/workdir/ExternalHeaders/{Library,StaticLibrary}
and fail.

Fix the incorrect gbuild usage and add some sanity checks to gbuild
to catch the problem closer to the source and more deterministically.