Force Linux builds to use gnu++98 (C++98 + GNU extensions) mode to be the most compatible with our old code base rather than relying on the compiler default mode.  Compiling in C++11 or newer mode is very noisy due to deprecation warnings about our use of std::auto_ptr.  If the compiler defaults to C++17 mode or newer, the build would be totally broken because std::auto_ptr is removed from C++17.
There is an unknown amount of porting effort needed to convert to
std::unique_ptr, which has somewhat different semantics and which is
not available before C++11, which would break building with older versions
of gcc which default to gnu++98 mode.

std::shared_ptr, might be an alternative since there is a BOOST
implementation.