Fix writerfilter build on CentOS 5 with recent libxslt.
xsltproc generates a corrupt writerfilter/.../inc/doctok/resources.hxx
file on CentOS 5 with libxslt 1.1.33 or newer.  This does not happen
with the CentOS 5 native version of xsltproc/libxslt, or with any version
of xsltproc/libxslt on CentOS 6 or newer.  The problem is that this xsl:
	<xsl:if test="$shift>0">
evaluates as true if $shift is an empty string on CentOS 5 with new
libxslt and evaluates as false everywhere else.  Work around this by
changing the xsl to:
	<xsl:if test="string-length($shift)>0 and $shift>0">
Curiously:
	<xsl:if test="$shift">
also does not work properly, which is surprising since both "0" and ""
should evaluate to false in a boolean context.