Fix for the Windows (MSI) installer.
The problem was that after 4.1.10 the productversion inside the msi was
"higher" than the one from AOO42X.

So for AOO 4.1.11 it is 4.111.9808. For AOO 4.2.0 it was 4.20.9823 (111 > 20).

This patch corrects the version number to 4.200.9823 (111 < 200).

Credits go to Arrigo for this patch!

(cherry picked from commit c1467797f39776453dad050af6d4bc996c2269b2)