Refs #i112383# and #i117960#. backporting to AOO41X .
Can't cherry-pick due to file changes.

Original commit messages and links follow:

Found-by: andrew
Patch-by: Damjan Jovanovic
https://github.com/apache/openoffice/commit/323c350123d05cd9f450aebd421b7bfa74c331d2

i92642 added the ability to use certain keywords as variable names (eg. name = 1, line = 'hi'),
but also caused a regression where "Line Input" is broken in single-line If statements.
This patch fixes that by allowing Then and Else to also be the start-of-line tokens expected to
immediately preceed the 'Line' token in order for that 'Line' token to be recognized a keyword instead
of a variable name. Also added FVT spreadsheet tests for 'Line' as both a variable name and as 'Line Input'.
Patch by: Damjan Jovanovic
https://github.com/apache/openoffice/commit/725d867363770a4c47c4b3c6dbe1f359c296604a