Minor cleanup to internal boost header packaging
We normally don't install all the headers availabe in boost. There is an
interesting set of functionalities like multiprecision and odeint that while we
are not using, but would be interesting to have available.

This is done to ease future boost updates but also in the hopes of making it
possible to use this functionality in a near future.

No functional change expected.