Allow our WebDAV content provider to connect when the TLS certificate name doesn't match the server's host name.
Currently in such cases the connection always fails, and the user isn't even
given a chance to allow it. This is because Curl does the server name
validation itself. However we already have code to validate server names,
and we prompt the user for what to do, unlike Curl which always fails.
Therefore disable Curl's verification and use ours.

Patch by: me