Implement the (MS Office 2010+) OOXML "Agile encryption" support, so that we can open such password-protected OOXML files.
Adds all the Agile encryption XML tokens and namespaces, and parses the XML
from EncryptionInfo stream, gets OpenOffice to recognize the file is encrypted
and ask for a password, and successfully decrypts the file if password is
correct.

Also a number of other fixes and improvements:
- Sorted main/oox/source/token/tokens.txt so it's in alphabetical order
  (wrong order might have broken certain tokens?).
- Refactored how OOXML encryption is generally handled. It's now in its
  own file.
- Added logging to the FilterDetect class. It logs to the office-wide default
  logger.
- Added a flush() method to the BinaryXOutputStream class.
- Changed FilterDetect to use XMultiComponentFactory and XComponentContext
  instead of the deprecated XMultiServiceFactory.
- Error handling was generally improved.
- Exception safety and some memory safety (::std::vector instead of new[])
  in all the new code. Memory leaks should not be possible.

Much of the code involved in the decryption was ported from the excellent
Apache POI project, so it's been credited in our NOTICE file.

Patch by: me