Add manual hook `chmod` to set file permissions for `md` files
chmod is not available on standard Windows so set this hook to manual.

Linux and macOS users can run this hook locally for a full repository clean up.

Also added the manual stage hook to GitHub Actions so we don't lose coverage

Example usage:

https://github.com/Lucas-C/pre-commit-hooks?tab=readme-ov-file#usage

https://github.com/apache/sedona/blob/3a1e9b7515b4dea994969f7f3c9ff43cf700e177/.pre-commit-config.yaml#L81

Finally renamed the workflow `linter.yml` to `pre-commit.yml`